// src/common/enums/messages.enum.ts
export enum messages {
    INVALID_ID_FORMAT = "Invalid Id format",
    UNAUTHORIZED_ACTION = "You are not authorized to perform this action",

    USER_NOT_FOUND = "User not found with this Id",
    USER_ID_REQUIRED = "User ID is required",

    POST_NOT_FOUND = "Post not found with this Id",
    DRAFT_POST_NOT_FOUND = "Draft post not found , either deleted or already published",
    POST_ID_REQUIRED = "Post ID is required",


    COMMENT_NOT_FOUND = "Comment not found with this Id",
    COMMENT_ID_REQUIRED = "Comment ID is required"
}